+++ plugin/src/main/java/KtlintConfigForm.kt
@@ -55,7 +55,14 @@ class KtlintConfigForm(private val project: Project, private val config: KtlintC
 
     fun createComponent(): JComponent {
         // Manually create and insert disabled rules field
-        disabledRules = TextFieldWithAutoCompletion.create(project, KtlintRules.getAll(config), false, "")
+        val rules = try {
+            KtlintRules.getAll(config)
+        } catch (ruleErr: Throwable) {
+            // UI for rule issues?
+            KtlintRules.getAll(config, true)
+        }
+
+        disabledRules = TextFieldWithAutoCompletion.create(project, rules, false, "")
         disabledRules.toolTipText = ResourceBundle.getBundle("strings").getString("disabledRulesToolTip")
         disabledRulesContainer.add(
             disabledRules,